decentralized config
